backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-pptx==0.6.23
lxml==5.2.2
anthropic==0.28.0
google-genai>=1.0.0
Pillow==10.3.0
scikit-image==0.23.2
scikit-learn==1.4.2
numpy==1.26.4
supabase==2.4.6
stripe==9.9.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
